Add support for snakemake > 8.0
