+++ snakebids/project_template/{% if build_system == 'poetry' %}pyproject.toml{% endif %}.jinja
@@ -22,13 +22,16 @@ classifiers = [
 
 [tool.poetry.dependencies]
 python = "{{ python_version }}"
-snakemake = ">={{ snakemake_version }},<8"
+snakemake = [
+    { version=">={{ snakemake_version }},<8", python = "<3.11" },
+    { version=">=8.1.2", python = ">=3.11" },
+]
 snakebids = {{ snakebids_version | format_poetry }}
 {#- newer pulps are incompatible with old snakemakes, and we need to support
 old snakemakes for python versions <3.11. So cap pulp to the last working
 version
 #}
-pulp = "<2.8.0"
+pulp = { version="<2.8.0", python = "<3.11" }
 pandas = [
     { version = "<=2.0.3", python = "<3.9" },
     { version = ">=2.1.1", python = ">=3.12" },
