+++ snakebids/tests/test_datasets.py
@@ -12,7 +12,6 @@ import more_itertools as itx
 import pytest
 from hypothesis import assume, given
 from hypothesis import strategies as st
-from snakemake.exceptions import WildcardError
 
 from snakebids.core.datasets import (
     BidsComponent,
@@ -22,6 +21,7 @@ from snakebids.core.datasets import (
 )
 from snakebids.exceptions import DuplicateComponentError
 from snakebids.paths._presets import bids
+from snakebids.snakemake_compat import WildcardError
 from snakebids.tests import strategies as sb_st
 from snakebids.tests.helpers import expand_zip_list, get_bids_path, get_zip_list, setify
 from snakebids.types import Expandable, ZipList
