--- typings/snakemake/exceptions.pyi
@@ -1,1 +1,0 @@
-class WildcardError(Exception): ...
